requirements.txt
python-dotenv~=0.21.0
httpx~=0.23.0